# openapi.json
{
  "components": {
    "schemas": {
...
      "CycleStateOut": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "current_node": {
            "title": "Current Node",
            "type": "string"
          },
          "cycle_id": {
            "title": "Cycle Id",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "transitions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/WorkflowTransitionOut"
            },
            "title": "Transitions",
            "type": "array"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "cycle_id",
          "current_node",
          "status"
        ],
        "title": "CycleStateOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunCycleIn": {
        "properties": {
          "force": {
            "default": false,
            "title": "Force",
            "type": "boolean"
          }
        },
        "title": "RunCycleIn",
        "type": "object"
      },
      "RunCycleOut": {
        "properties": {
          "cycle_id": {
            "title": "Cycle Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "cycle_id",
          "status",
          "message",
          "started_at"
        ],
        "title": "RunCycleOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkflowStateOut": {
        "description": "Current workflow state snapshot for one cycle.",
        "properties": {
          "current_node": {
            "title": "Current Node",
            "type": "string"
          },
          "cycle_id": {
            "title": "Cycle Id",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "cycle_id",
          "current_node",
          "status"
        ],
        "title": "WorkflowStateOut",
        "type": "object"
      },
      "WorkflowTransitionOut": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "from_node": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Node"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "to_node": {
            "title": "To Node",
            "type": "string"
          },
          "transitioned_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transitioned At"
          }
        },
        "required": [
          "to_node",
          "status"
        ],
        "title": "WorkflowTransitionOut",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Autonomous Adaptive Portfolio Hedge Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/cycle/{cycle_id}/state": {
      "get": {
        "description": "Return current node, status, and transition history for a cycle.",
        "operationId": "get_cycle_state_cycle__cycle_id__state_get",
        "parameters": [
          {
            "description": "Unique ID of the orchestrator run cycle",
            "in": "path",
            "name": "cycle_id",
            "required": true,
            "schema": {
              "description": "Unique ID of the orchestrator run cycle",
              "title": "Cycle Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleStateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Cycle State",
        "tags": [
          "workflow"
        ]
      }
    },
...
    "/run-cycle": {
      "post": {
        "description": "Trigger an autonomous hedge assessment cycle (P6-BE-11 / #157).\n\nStarts the six-node LangGraph pipeline in a background thread and returns\n202 Accepted immediately with the new cycle_id.  Progress is readable via\nGET /cycle/{cycle_id}/state or GET /workflow-state/stream?cycle_id=...",
        "operationId": "trigger_run_cycle_run_cycle_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCycleIn",
                "default": {
                  "force": false
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCycleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trigger Run Cycle",
        "tags": [
          "workflow"
        ]
      }
    },
...
    "/workflow-state": {
      "get": {
        "description": "Return the current workflow state (P6-BE-12 / #158).\n\ncycle_id is optional; omitting it returns the most-recently updated cycle.\nReturns 404 when no state has been persisted yet.",
        "operationId": "get_workflow_state_workflow_state_get",
        "parameters": [
          {
            "description": "Cycle to inspect; omit for latest",
            "in": "query",
            "name": "cycle_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cycle to inspect; omit for latest",
              "title": "Cycle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workflow State",
        "tags": [
          "workflow"
        ]
      }
    },
    "/workflow-state/stream": {
      "get": {
        "description": "SSE stream of workflow-state transitions (P6-BE-12 / #158).\n\nEmits one ``data: <JSON>`` Server-Sent Event per new row in\n``workflow_transitions`` for the requested cycle (or the most-recently\nactive cycle when cycle_id is omitted).\n\nEach event body is a JSON object with keys cycle_id, node, status, ts.\n\nThe stream closes automatically when:\n\n* the cycle reaches a terminal node (COMPLETED / FAILED); or\n* the MONITORING EXIT transition is emitted.",
        "operationId": "stream_workflow_state_workflow_state_stream_get",
        "parameters": [
          {
            "description": "Cycle to stream; omit for latest",
            "in": "query",
            "name": "cycle_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cycle to stream; omit for latest",
              "title": "Cycle Id"
            }
          },
          {
            "in": "query",
            "name": "poll_interval_s",
            "required": false,
            "schema": {
              "default": 0.5,
              "maximum": 10.0,
              "minimum": 0.0,
              "title": "Poll Interval S",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Stream Workflow State Workflow State Stream Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Workflow State",
        "tags": [
          "workflow"
        ]
      }
    },
    "/workflow/state": {
      "get": {
        "description": "Return the current workflow state (P6-BE-12 / #158).\n\ncycle_id is optional; omitting it returns the most-recently updated cycle.\nReturns 404 when no state has been persisted yet.",
        "operationId": "get_workflow_state_workflow_state_get",
        "parameters": [
          {
            "description": "Cycle to inspect; omit for latest",
            "in": "query",
            "name": "cycle_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cycle to inspect; omit for latest",
              "title": "Cycle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workflow State",
        "tags": [
          "workflow"
        ]
      }
    },
    "/workflow/state/stream": {
      "get": {
        "description": "SSE stream of workflow-state transitions (P6-BE-12 / #158).\n\nEmits one ``data: <JSON>`` Server-Sent Event per new row in\n``workflow_transitions`` for the requested cycle (or the most-recently\nactive cycle when cycle_id is omitted).\n\nEach event body is a JSON object with keys cycle_id, node, status, ts.\n\nThe stream closes automatically when:\n\n* the cycle reaches a terminal node (COMPLETED / FAILED); or\n* the MONITORING EXIT transition is emitted.",
        "operationId": "stream_workflow_state_workflow_state_stream_get",
        "parameters": [
          {
            "description": "Cycle to stream; omit for latest",
            "in": "query",
            "name": "cycle_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cycle to stream; omit for latest",
              "title": "Cycle Id"
            }
          },
          {
            "in": "query",
            "name": "poll_interval_s",
            "required": false,
            "schema": {
              "default": 0.5,
              "maximum": 10.0,
              "minimum": 0.0,
              "title": "Poll Interval S",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Stream Workflow State Workflow State Stream Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Workflow State",
        "tags": [
          "workflow"
        ]
      }
    }
  }
}